[#9] Move oauth endpoint to sub api/v1
